:pencil2: 플레이리스트 template 추가, song template 분리
