+++ src/templates/search.ts
@@ -1,19 +1,4 @@
-export type Song = {
-  songId: string;
-  title: string;
-  artist: string;
-  remix: string;
-  reaction: string;
-  date: number;
-  start: number;
-  end: number;
-  total: {
-    views: number | null;
-    increase: number | null;
-    last: number | null;
-  };
-  like: number;
-};
+import { Song } from "./song";
 
 export type SongsSearchResponse = {
   song: Song[];
