+++ src/templates/song.ts
@@ -1,0 +1,16 @@
+export type Song = {
+  songId: string;
+  title: string;
+  artist: string;
+  remix: string;
+  reaction: string;
+  date: number;
+  start: number;
+  end: number;
+  total: {
+    views: number | null;
+    increase: number | null;
+    last: number | null;
+  };
+  like?: number;
+};
