+++ src/templates/user.ts
@@ -1,0 +1,7 @@
+export type User = {
+  displayName: string;
+  profile: {
+    type: string;
+    version: number;
+  };
+};
